beams-sandbox/workspace/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langsmith>=0.7.4",
    "openai>=2.21.0",
    "tavily-python>=0.7.21",
]
